Fixed tests with new custom exceptions. Custom exceptions restructured more logically

+++ include/FiniteElementMethod/Cubature/BasisSelector.hpp
@@ -11,7 +11,7 @@
 #include <memory>
 
 #include "FiniteElementMethod/Cell/CellType.hpp"
-#include "FiniteElementMethod/Cubature/CubatureExceptions.hpp"
+#include "FiniteElementMethod/FEMExceptions.hpp"
 #include "FiniteElementMethod/Utils/FEMCheckers.hpp"
 #include "FiniteElementMethod/FEMTypes.hpp"
 
@@ -43,7 +43,7 @@ public:
      * @throw BasisSelectorUnsupportedCellTypeException If the cell type is not supported.
      * @throw BasisSelectorUnsupportedPolynomOrderException If the polynomial order is not supported.
      */
-    static std::unique_ptr<Intrepid2::Basis<DeviceType>> get(CellType cellType, int polynom_order)
+    static std::unique_ptr<Intrepid2::Basis<DeviceType>> get(CellType cellType, short polynom_order)
     {
         FEMCheckers::checkPolynomOrder(polynom_order);
